Refactor TextField for merge styles work
