+++ packages/office-ui-fabric-react/src/components/TextField/TextField.styles.ts
@@ -1,0 +1,9 @@
+import { ITextFieldStyleProps, ITextFieldStyles } from '../..';
+
+const GlobalClassNames = {
+
+};
+
+export const getStyles = (props: ITextFieldStyleProps): ITextFieldStyles => {
+  return ({});
+};
