+++ packages/office-ui-fabric-react/src/components/TextField/TextField.types.ts
@@ -1,6 +1,7 @@
 import * as React from 'react';
-import { IRenderFunction } from '../../Utilities';
+import { IRenderFunction, IStyleFunction } from '../../Utilities';
 import { IIconProps } from '../../Icon';
+import { ITheme } from '../..';
 
 export interface ITextField {
   /** Gets the current value of the input. */
@@ -41,6 +42,16 @@ export interface ITextFieldProps extends React.AllHTMLAttributes<HTMLInputElemen
    * the public methods and properties of the component.
    */
   componentRef?: (component: ITextField | null) => void;
+
+  /**
+   * Call to provide customized styling that will layer on top of the variant rules.
+   */
+  getStyles?: IStyleFunction<ITextFieldStyleProps, ITextFieldStyles>;
+
+  /**
+   * Theme to apply to the component.
+   */
+  theme?: ITheme;
 
   /**
    * Whether or not the textfield is a multiline textfield.
@@ -264,4 +275,12 @@ export interface ITextFieldProps extends React.AllHTMLAttributes<HTMLInputElemen
    * @deprecated
    */
   componentId?: string;
+}
+
+export interface ITextFieldStyleProps {
+
+}
+
+export interface ITextFieldStyles {
+
 }
