+++ packages/office-ui-fabric-react/src/components/TextField/index.ts
@@ -1,3 +1,4 @@
 export * from './TextField';
+export * from './TextField.base';
 export * from './TextField.types';
 export * from './MaskedTextField/MaskedTextField';
